INP-Former/requirements.txt
adeval==1.1.0
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
cmake==4.2.3
coloredlogs==15.0.1
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.124.4
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==25.12.19
fsspec==2025.3.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
imageio==2.35.1
importlib_resources==6.4.5
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.10
lit==18.1.8
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.2.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.19.5
onnx==1.15.0
onnxruntime-gpu==1.15.0
onnxsim==0.4.33
opencv-python-headless==4.6.0.66
opencv-python==4.13.0.92
orjson==3.10.15
packaging==26.0
pandas==1.3.5
pillow==9.0.1
protobuf==5.29.6
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
pywavelets==1.4.1
pyyaml==6.0.3
requests==2.32.4
rich==14.3.3
ruff==0.15.5
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
tifffile==2023.7.10
timm==0.9.12
tomlkit==0.12.0
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.64.1
triton==2.0.0
typer==0.20.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
websockets==12.0
zipp==3.20.2
